requirements.txt
django
django-filter
djangorestframework
djangorestframework-gis
psycopg2-binary
